agent/sub-agents/backend/requirements.txt
# Core dependencies
google-adk>=1.0.0
google-genai>=0.3.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.28.0          # PostgreSQL async driver
aiomysql>=0.2.0          # MySQL async driver
psycopg2-binary>=2.9.0   # PostgreSQL sync driver
pymysql>=1.1.0           # MySQL sync driver

# HTTP client dependencies
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0

# Authentication dependencies
pyjwt>=2.8.0
bcrypt>=4.0.0
cryptography>=41.0.0
python-multipart>=0.0.6

# File processing dependencies
pandas>=2.0.0
openpyxl>=3.1.0          # Excel files
PyPDF2>=3.0.0            # PDF processing
python-docx>=0.8.11      # Word documents
Pillow>=10.0.0           # Image processing

# Data format dependencies
xmltodict>=0.13.0        # XML processing
python-magic>=0.4.27     # File type detection
chardet>=5.2.0           # Character encoding detection

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0            # Code formatting
isort>=5.12.0            # Import sorting
flake8>=6.0.0            # Linting
mypy>=1.5.0              # Type checking

# Monitoring and logging
structlog>=23.1.0        # Structured logging
prometheus-client>=0.17.0 # Metrics
sentry-sdk>=1.32.0       # Error tracking

# Utility dependencies
python-dotenv>=1.0.0     # Environment variables
click>=8.1.0             # CLI interface
rich>=13.5.0             # Rich console output
backoff>=2.2.0           # Retry with backoff

# Optional cloud dependencies
google-cloud-storage>=2.10.0    # Google Cloud Storage
boto3>=1.28.0                   # AWS SDK
azure-storage-blob>=12.17.0     # Azure Blob Storage

# Optional message queue dependencies
redis>=4.6.0             # Redis client
celery>=5.3.0            # Task queue
kombu>=5.3.0             # Message broker

# Optional monitoring dependencies
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation>=0.40b0

# Security dependencies
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4             # Password hashing

# Data validation and serialization
marshmallow>=3.20.0      # Schema validation
cerberus>=1.3.4          # Data validation

# Async utilities
anyio>=3.7.0             # Async utilities
trio>=0.22.0             # Alternative async library

# Configuration management
dynaconf>=3.2.0          # Configuration management
python-decouple>=3.8     # Settings from environment

# API documentation
swagger-ui-bundle>=0.0.9  # Swagger UI
redoc>=0.0.4             # ReDoc documentation

# Development tools
pre-commit>=3.3.0        # Git hooks
coverage>=7.3.0          # Code coverage
bandit>=1.7.5            # Security linting
safety>=2.3.0            # Security vulnerability scanner

# Database migration tools
alembic>=1.12.0          # Database migrations

# Container support
gunicorn>=21.2.0         # WSGI server
docker>=6.1.0            # Docker client

# Caching
cachetools>=5.3.0        # In-memory caching
diskcache>=5.6.0         # Disk-based caching

# Serialization formats
msgpack>=1.0.0           # MessagePack serialization
cbor2>=5.4.0             # CBOR serialization

# Network utilities
dnspython>=2.4.0         # DNS utilities
netaddr>=0.8.0           # Network address manipulation

# Text processing
textdistance>=4.6.0      # Text similarity
fuzzywuzzy>=0.18.0       # Fuzzy string matching

# Time utilities
python-dateutil>=2.8.0   # Date utilities
pytz>=2023.3             # Timezone handling

# Validation utilities
email-validator>=2.0.0   # Email validation
phonenumbers>=8.13.0     # Phone number validation

# Template engines
jinja2>=3.1.0            # Template engine for reports

# Excel advanced processing
xlsxwriter>=3.1.0        # Excel writing
xlrd>=2.0.0              # Excel reading (legacy)

# PDF advanced processing
reportlab>=4.0.0         # PDF generation
pdfminer.six>=20221105   # PDF text extraction

# Image processing advanced
opencv-python>=4.8.0     # Computer vision
scikit-image>=0.21.0     # Image processing algorithms

# Machine learning (optional)
scikit-learn>=1.3.0      # Machine learning
numpy>=1.24.0            # Numerical computing
scipy>=1.11.0            # Scientific computing

# Natural language processing (optional)
nltk>=3.8.0              # Natural language toolkit
spacy>=3.6.0             # Advanced NLP

# Workflow management
prefect>=2.13.0          # Workflow orchestration
dask>=2023.8.0           # Parallel computing

# API rate limiting
slowapi>=0.1.9           # Rate limiting for FastAPI
limits>=3.5.0            # Rate limiting utilities

# WebSocket support
websockets>=11.0.0       # WebSocket client/server
starlette>=0.27.0        # ASGI framework

# GraphQL support (optional)
graphene>=3.3.0          # GraphQL framework
strawberry-graphql>=0.200.0  # Modern GraphQL

# Geospatial support (optional)
geoalchemy2>=0.14.0      # Geospatial database extensions
shapely>=2.0.0           # Geometric operations

# Financial data processing (optional)
python-forex>=1.8       # Currency conversion
yfinance>=0.2.0          # Financial data

# Social media APIs (optional)
tweepy>=4.14.0           # Twitter API
facebook-sdk>=3.1.0      # Facebook API

# Notification services (optional)
plyer>=2.1.0             # Cross-platform notifications
pushbullet.py>=0.12.0    # Pushbullet notifications

